requirements.txt
requests
python-dotenv
gradio
pypdf
openai
openai-agents
apify-client
pygithub
chromadb
langchain
langchain-openai
langchain-community